remove an unneeded cast

+++ state_tracker/state_bufferobject.c
@@ -348,7 +348,7 @@ crStateBufferSubDataARB(GLenum target, GLintptrARB offset, GLsizeiptrARB size, c
 		return;
 	}
 
-	if (size < 0 || (unsigned int)offset < 0 || (unsigned int)offset + size > obj->size) {
+	if (size < 0 || offset < 0 || (unsigned int)offset + size > obj->size) {
 		crStateError(__LINE__, __FILE__, GL_INVALID_OPERATION,
 								 "glBufferSubDataARB(bad offset and/or size)");
 		return;
@@ -407,7 +407,7 @@ crStateGetBufferSubDataARB(GLenum target, GLintptrARB offset, GLsizeiptrARB size
 		return;
 	}
 
-	if (size < 0 || (unsigned int)offset < 0 || (unsigned int)offset + size > obj->size) {
+	if (size < 0 || offset < 0 || (unsigned int)offset + size > obj->size) {
 		crStateError(__LINE__, __FILE__, GL_INVALID_OPERATION,
 								 "glGetBufferSubDataARB(bad offset and/or size)");
 		return;
